component: GenericDropdown moved inside hexagon
