+++ src/components/DatabaseTables/VocabQuizDb/StudentsTable/components/FilterStudentsTable.tsx
@@ -1,8 +1,11 @@
 import type { Dispatch, SetStateAction } from 'react';
 import type { FilterConfig } from 'src/components/Table/types';
-import { Dropdown } from '@interface/components/FormComponents';
+import {
+  Dropdown,
+  GenericDropdown,
+} from '@interface/components/FormComponents';
 import { useEffect, useState } from 'react';
-import { GenericDropdown, TextInput } from 'src/components/FormComponents';
+import { TextInput } from 'src/components/FormComponents';
 import { useContextualMenu } from 'src/hexagon/interface/hooks/useContextualMenu';
 import useStudentsTable from '../useStudentsTable';
