+++ src/hexagon/interface/components/FormComponents/index.ts
@@ -1,3 +1,4 @@
+import type { GenericDropdownOption } from '@interface/components/FormComponents/Dropdown/GenericDropdown';
 import Dropdown from '@interface/components/FormComponents/Dropdown/Dropdown';
-
-export { Dropdown };
+import GenericDropdown from '@interface/components/FormComponents/Dropdown/GenericDropdown';
+export { Dropdown, GenericDropdown, type GenericDropdownOption };
